Some file structure setup and file_compare module
Some initial file structure setup including a package called files which will contain the modules used for file checking/searching/etc. In the files package there is a tests subpackage which currently contains an initial_test for the file_compare module. This module contains CRC32 hashing algorithm implementation as well as some additional methods for reading files to certain data structures as well as a method which can check whether two files are equal. This will be the basis for the version control file checking algorithm.