More functionality added to command_handler. Moved forkie.py to root dir
Moved forkie.py to root directory so that when it called the command_handler functions it would have the path initialised to access other packages such as src. Changed the docopt inside forkie.py to fix some logical errors. Added a command to open a temporary message (in command_handler) inside the default editor when command is missing message option.